Remove base_classes and envopt=None
